test(protocol): add comprehensive encode/decode unit tests
Add 45 new tests covering all protocol encode/decode functions:
- Passphrase, Scan, Connect, CreateAccessPoint
- SetAcceptPolicy, SetAdvertiseData, ProxyData
- ProxyConnect, ProxyConnectReply, ProxyDisconnect
- Reject, NetworkInfo, ScanReplyEnd, RejectReply
- Raw data, encode_with_data, utility functions
- All packet_id_to_string, result_to_string values
- Additional structure size validations
- Error handling edge cases

Fix encode_set_accept_policy to use uint8_t cast (matches protocol).

+++ sysmodule/source/protocol/ryu_protocol.hpp
@@ -282,7 +282,7 @@ inline EncodeResult encode_create_access_point(uint8_t* buffer, size_t buffer_si
 inline EncodeResult encode_set_accept_policy(uint8_t* buffer, size_t buffer_size,
                                              AcceptPolicy policy, size_t& out_size) {
     SetAcceptPolicyRequest msg{};
-    msg.accept_policy = static_cast<uint32_t>(policy);
+    msg.accept_policy = static_cast<uint8_t>(policy);
     return encode(buffer, buffer_size, PacketId::SetAcceptPolicy, msg, out_size);
 }
